Taping Out:
Right now all the code I've been working with in regards to orientation detection is commented out. I got to the point where the program recongizines the oriteaton of a screen on a desktop screen but am struggling to get the program to recongize the screen orienation on a mobile device.